+++ src/config/types/index.ts
@@ -28,7 +28,7 @@ export type Configuration = Readonly<
     maxNamesPerWrappedLine: number;
     keepUnused: KeepUnusedRule[];
     EmptyLinesBetweenGroups: number;
-    EmptyLinesAfterAllImports: number;
+    emptyLinesAfterAllImports: number;
     // From other configs
     maxLineLength: number;
     tabType: 'space' | 'tab';