+++ src/format/config/compose.ts
@@ -11,7 +11,7 @@ export function configForCompose({
   maxExportNamesPerLine,
   maxNamesPerWrappedLine,
   EmptyLinesBetweenGroups,
-  EmptyLinesAfterAllImports,
+  emptyLinesAfterAllImports,
   tabType,
   tabSize,
   quoteMark,
@@ -31,7 +31,7 @@ export function configForCompose({
       exported: maxExportNamesPerLine || Number.MAX_SAFE_INTEGER,
     },
     groupSep: nl.repeat((EmptyLinesBetweenGroups ?? 1) + 1),
-    groupEnd: (EmptyLinesAfterAllImports ?? 1) + 1,
+    groupEnd: (emptyLinesAfterAllImports ?? 1) + 1,
     tab: tabType?.toLowerCase() === 'tab' ? '\t' : ' '.repeat(tabSize ?? 2),
     quote:
       quoteMark?.toLowerCase() === 'double' ? (s: string) => `"${s}"` : (s: string) => `'${s}'`,