+++ src/test/suite/extension.test.ts
@@ -46,7 +46,7 @@ suite('Integration Test Suite', () => {
   // Run all tests
   return runTestSuite(examples);
   // Or, run a specific test case
-  // return runTestSuite(examples, 'compose/group-end');
+  // return runTestSuite(examples, 'eslint/import');
 });
 
 function getTestSuite(dir: string, name: string): TestSuite | undefined {